fix: Extend tests to check ExportRun and move import statement
